Part3/Chapter02-03-image-filter/requirements.txt
streamlit==1.44.1
numpy==2.2.5
pillow==11.2.1